+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 
 ### Added
 
-- You can now configure plugins on each client in `extra_plugins`. Plugins that you previously had to define your own services can be configured on the client.
+- You can now also configure client specific plugins in the `plugins` option of a client.
+  Some plugins that you previously had to define in your own services can now be configured on the client.
 - Support for BatchClient
 - The stopwatch plugin in included by default when using profiling.